pat_re_eval.t: remove 'no warnings'
Now the the closure behaviour of code blocks within regexes is mostly
fixes, remove the

    no warnings qw(uninitialized closure);

that protected many of the closure tests.

This also showed up that one of my recent tests didn't realise
the whose test file is wrapped in a sub, so my test code destructor
looked like

    sub run_tests {
	...
	my $d;
	sub DESTRUCTOR { $d++ }
	...
    }

so fix that test to remove the closure warning.