Move stack extension into FT_* macros
Every filetest operator with the OPf_REF flag set has to extend the
stack for the argument.  One operator was missing it until the previ-
ous commit.  Since we have special return macros for these operators,
put the stack extension there, to avoid a repeat of this type of mis-
take and reduce repetitiveness.