regen/opcodes: Rmv evalonce comment
This goes all the way back to when perl 5.0 was being polished up.
The idea behind evalonce is that eval "constant string" should be
optimisable by being compiled ahead of time, just like eval { ... }.
But this could never work properly, because BEGIN blocks would only
fire once.  Also, eval '$x .. $y' is an easy way to create two separ-
ate flip-flops.  There are undoubtedly many other reasons why this
could never work.

So there is no reason to keep this comment any longer, as it is not
(or should not be) a to-do item.