scope.c: Simplify and clarify comment
This comment seems to imply that this code is just working around a
problem in gv.c, which we could simply correct there.  I’ve already
tried making the quoted code in gv.c handle *^H without a hash, but it
doesn’t actually solve the problem.  The real problem is that rv2hv
could also add a hash to *^H, via GvHVn, and that is simply not set up
to deal with autovivifying magic at all, and probably shouldn’t be.

Also, quoting a piece of code that occurs elsewhere is just asking for
it to drift apart.  By this time, the code in gv.c that is quoted in
scope.c is actually three times the length now, and looks completely
different.