Conditionally compile QNX/NTO code in File::Spec::Unix::canonpath()
File::Spec::Unix::canonpath() does a platform run-time check at each
invocation to apply a QNX/NTO hack. As this is not needed on plain Unix
systems, the check can be moved to a constant so that the special code
can be optimized away by constant folded.

This seems to yield a 25% speedup for each call to canonpath().